+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/reviewtransaction/AndroidReviewTransaction.kt
@@ -1,15 +1,14 @@
-package co.electriccoin.zcash.ui.screen.transactionprogress
+package co.electriccoin.zcash.ui.screen.reviewtransaction
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import org.koin.androidx.compose.koinViewModel
-import org.koin.core.parameter.parametersOf
 
 @Composable
-fun AndroidKeystoneTransactionProgress(args: KeystoneTransactionProgress) {
-    val viewModel = koinViewModel<KeystoneTransactionProgressViewModel> { parametersOf(args) }
+fun AndroidReviewTransaction() {
+    val viewModel = koinViewModel<ReviewTransactionViewModel>()
     val state by viewModel.state.collectAsStateWithLifecycle()
 
     BackHandler {
@@ -17,6 +16,6 @@ fun AndroidKeystoneTransactionProgress(args: KeystoneTransactionProgress) {
     }
 
     state?.let {
-        TransactionProgressView(it)
+        ReviewTransactionView(it)
     }
 }