+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/reviewtransaction/ReviewTransaction.kt
@@ -3,4 +3,4 @@ package co.electriccoin.zcash.ui.screen.reviewtransaction
 import kotlinx.serialization.Serializable
 
 @Serializable
-data object ReviewKeystoneTransaction
+data object ReviewTransaction