+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/send/SendViewModel.kt
@@ -2,21 +2,12 @@ package co.electriccoin.zcash.ui.screen.send
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import cash.z.ecc.android.sdk.ext.convertZecToZatoshi
-import cash.z.ecc.android.sdk.model.Memo
-import cash.z.ecc.android.sdk.model.WalletAddress
 import cash.z.ecc.android.sdk.model.ZecSend
-import cash.z.ecc.android.sdk.model.proposeSend
-import cash.z.ecc.android.sdk.type.AddressType
 import cash.z.ecc.sdk.ANDROID_STATE_FLOW_TIMEOUT
 import co.electriccoin.zcash.spackle.Twig
-import co.electriccoin.zcash.ui.common.model.KeystoneAccount
-import co.electriccoin.zcash.ui.common.model.ZashiAccount
-import co.electriccoin.zcash.ui.common.usecase.CreateKeystoneProposalUseCase
-import co.electriccoin.zcash.ui.common.usecase.CreateKeystoneZip321ProposalUseCase
-import co.electriccoin.zcash.ui.common.usecase.GetSelectedWalletAccountUseCase
-import co.electriccoin.zcash.ui.common.usecase.GetSynchronizerUseCase
-import co.electriccoin.zcash.ui.common.usecase.GetZashiAccountUseCase
+import co.electriccoin.zcash.ui.common.usecase.CreateProposalUseCase
+import co.electriccoin.zcash.ui.common.usecase.CreateZip321ProposalUseCase
+import co.electriccoin.zcash.ui.common.usecase.NavigateToAddressBookUseCase
 import co.electriccoin.zcash.ui.common.usecase.ObserveContactByAddressUseCase
 import co.electriccoin.zcash.ui.common.usecase.ObserveContactPickedUseCase
 import co.electriccoin.zcash.ui.common.usecase.ObserveWalletAccountsUseCase
@@ -37,18 +28,15 @@ import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import org.zecdev.zip321.ZIP321
 import kotlin.time.Duration.Companion.seconds
 
 class SendViewModel(
     private val observeContactByAddress: ObserveContactByAddressUseCase,
     private val observeContactPicked: ObserveContactPickedUseCase,
-    private val getSynchronizer: GetSynchronizerUseCase,
-    private val getSelectedWalletAccount: GetSelectedWalletAccountUseCase,
-    private val getZashiAccount: GetZashiAccountUseCase,
-    private val createKeystoneTransactionProposal: CreateKeystoneProposalUseCase,
-    private val createKeystoneZip321TransactionProposal: CreateKeystoneZip321ProposalUseCase,
+    private val createProposal: CreateProposalUseCase,
+    private val createKeystoneZip321TransactionProposal: CreateZip321ProposalUseCase,
     private val observeWalletAccounts: ObserveWalletAccountsUseCase,
+    private val navigateToAddressBook: NavigateToAddressBookUseCase
 ) : ViewModel() {
     val recipientAddressState = MutableStateFlow(RecipientAddressState.new("", null))
 
@@ -128,7 +116,7 @@ class SendViewModel(
         when (mode) {
             SendAddressBookState.Mode.PICK_FROM_ADDRESS_BOOK ->
                 viewModelScope.launch {
-                    navigateCommand.emit(AddressBookArgs(AddressBookArgs.PICK_CONTACT))
+                    navigateToAddressBook(AddressBookArgs.PICK_CONTACT)
                 }
 
             SendAddressBookState.Mode.ADD_TO_ADDRESS_BOOK ->
@@ -142,124 +130,29 @@ class SendViewModel(
         recipientAddressState.update { state }
     }
 
+    @Suppress("TooGenericExceptionCaught")
     fun onCreateZecSendClick(
         newZecSend: ZecSend,
-        setZecSend: (ZecSend) -> Unit,
-        goSendConfirmation: (ZecSend) -> Unit,
         setSendStage: (SendStage) -> Unit
     ) = viewModelScope.launch {
-        when (getSelectedWalletAccount()) {
-            is KeystoneAccount -> {
-                try {
-                    createKeystoneTransactionProposal(newZecSend)
-                } catch (_: Exception) {
-                    setSendStage(SendStage.SendFailure(""))
-                }
-            }
-            is ZashiAccount -> {
-                Twig.debug { "Getting send transaction proposal" }
-                runCatching {
-                    getSynchronizer().proposeSend(getZashiAccount().sdkAccount, newZecSend)
-                }.onSuccess { proposal ->
-                    Twig.debug { "Transaction proposal successful: ${proposal.toPrettyString()}" }
-                    val enrichedZecSend = newZecSend.copy(proposal = proposal)
-                    setZecSend(enrichedZecSend)
-                    goSendConfirmation(enrichedZecSend)
-                }.onFailure {
-                    Twig.error(it) { "Transaction proposal failed" }
-                    setSendStage(SendStage.SendFailure(it.message ?: ""))
-                }
-            }
+        try {
+            createProposal(newZecSend)
+        } catch (e: Exception) {
+            setSendStage(SendStage.SendFailure(""))
+            Twig.error(e) { "Error creating proposal" }
         }
     }
 
+    @Suppress("TooGenericExceptionCaught")
     fun onCreateZecSend321Click(
         zip321Uri: String,
-        setZecSend: (ZecSend) -> Unit,
         setSendStage: (SendStage) -> Unit,
-        goPaymentRequest: (ZecSend, String) -> Unit,
     ) = viewModelScope.launch {
-        when (getSelectedWalletAccount()) {
-            is KeystoneAccount -> {
-                try {
-                    createKeystoneZip321TransactionProposal(zip321Uri)
-                } catch (_: Exception) {
-                    setSendStage(SendStage.SendFailure(""))
-                }
-            }
-            is ZashiAccount ->
-                onCreateZecSend321ZashiClick(
-                    zip321Uri = zip321Uri,
-                    setZecSend = setZecSend,
-                    setSendStage = setSendStage,
-                    goPaymentRequest = goPaymentRequest,
-                )
+        try {
+            createKeystoneZip321TransactionProposal(zip321Uri)
+        } catch (e: Exception) {
+            setSendStage(SendStage.SendFailure(""))
+            Twig.error(e) { "Error creating proposal" }
         }
     }
-
-    private suspend fun onCreateZecSend321ZashiClick(
-        zip321Uri: String,
-        setZecSend: (ZecSend) -> Unit,
-        setSendStage: (SendStage) -> Unit,
-        goPaymentRequest: (ZecSend, String) -> Unit,
-    ) {
-        val synchronizer = getSynchronizer()
-        val account = getZashiAccount()
-
-        val request =
-            runCatching {
-                // At this point there should by only a valid Zcash address coming
-                ZIP321.request(zip321Uri, null)
-            }.onFailure {
-                Twig.error(it) { "Failed to validate address" }
-            }.getOrElse {
-                false
-            }
-        val payment =
-            when (request) {
-                // We support only one payment currently
-                is ZIP321.ParserResult.Request -> {
-                    request.paymentRequest.payments[0]
-                }
-                else -> return
-            }
-
-        val address =
-            synchronizer
-                .validateAddress(payment.recipientAddress.value)
-                .toWalletAddress(payment.recipientAddress.value)
-
-        val amount = payment.nonNegativeAmount.value.convertZecToZatoshi()
-
-        val memo = Memo(payment.memo?.let { String(it.data, Charsets.UTF_8) } ?: "")
-
-        val zecSend =
-            ZecSend(
-                destination = address,
-                amount = amount,
-                memo = memo,
-                proposal = null
-            )
-        setZecSend(zecSend)
-
-        runCatching {
-            synchronizer.proposeFulfillingPaymentUri(account.sdkAccount, zip321Uri)
-        }.onSuccess { proposal ->
-            Twig.debug { "Transaction proposal from Zip321 Uri: ${proposal.toPrettyString()}" }
-            val enrichedZecSend = zecSend.copy(proposal = proposal)
-            setZecSend(enrichedZecSend)
-            goPaymentRequest(enrichedZecSend, zip321Uri)
-        }.onFailure {
-            Twig.error(it) { "Transaction proposal from Zip321 Uri failed" }
-            setSendStage(SendStage.SendFailure(it.message ?: ""))
-        }
-    }
-
-    private suspend fun AddressType.toWalletAddress(value: String) =
-        when (this) {
-            AddressType.Unified -> WalletAddress.Unified.new(value)
-            AddressType.Shielded -> WalletAddress.Sapling.new(value)
-            AddressType.Transparent -> WalletAddress.Transparent.new(value)
-            else -> error("Invalid address type")
-        }
 }